fix(fleet): 阻止刷新标记进入常驻进程
